move barcode calling helpers out of isoquant.py
