+++ isoquant_lib/barcode_calling/__init__.py
@@ -76,6 +76,15 @@ from .detect_barcodes import (
 # Cell barcode detection
 from .cell_selection import CellBarcodeSelector, estimate_cell_number, select_cell_barcodes
 
+# Command line resolution and the pipeline stage
+from .options import (
+    resolve_barcode_correction,
+    resolve_deprecated_mode,
+    resolve_split_molecules,
+    validate_barcode_calling,
+)
+from .pipeline import BarcodeCallingArgs, call_barcodes, detect_cell_barcodes
+
 __all__ = [
     # Indexers
     'KmerIndexer',
@@ -131,4 +140,12 @@ __all__ = [
     'CellBarcodeSelector',
     'estimate_cell_number',
     'select_cell_barcodes',
+    # Command line resolution and the pipeline stage
+    'resolve_barcode_correction',
+    'resolve_deprecated_mode',
+    'resolve_split_molecules',
+    'validate_barcode_calling',
+    'BarcodeCallingArgs',
+    'call_barcodes',
+    'detect_cell_barcodes',
 ]
