+++ isoquant_tests/test_cell_barcode_options.py
@@ -10,7 +10,7 @@ import argparse
 
 import pytest
 
-import isoquant
+from isoquant_lib.barcode_calling import options
 from isoquant_lib.modes import AUTO_BARCODES, BarcodeCorrectionMethod, IsoQuantMode
 
 
@@ -34,37 +34,37 @@ class TestResolveNCells:
                                                  ("5000", 5000)])
     def test_accepted_values(self, value, expected):
         args = make_args(n_cells=value)
-        assert isoquant._resolve_n_cells(args) == expected
+        assert options.resolve_n_cells(args) == expected
 
     @pytest.mark.parametrize("value", ["0", "-1", "many", "1e3"])
     def test_rejected_values(self, value):
         with pytest.raises(SystemExit) as excinfo:
-            isoquant._resolve_n_cells(make_args(n_cells=value))
+            options.resolve_n_cells(make_args(n_cells=value))
         assert excinfo.value.code != 0
 
 
 class TestDetectionIsRequested:
     def test_whitelist_alone_is_the_cell_list(self, tmp_path):
         args = make_args(barcode_whitelist=whitelist_file(tmp_path))
-        isoquant._resolve_barcode_correction(args)
+        options.resolve_barcode_correction(args)
         assert args.detect_cell_barcodes is False
 
     def test_n_cells_turns_the_whitelist_into_a_pool(self, tmp_path):
         args = make_args(n_cells="5000", barcode_whitelist=whitelist_file(tmp_path))
-        isoquant._resolve_barcode_correction(args)
+        options.resolve_barcode_correction(args)
         assert args.detect_cell_barcodes is True
         assert args.n_cells == 5000
 
     def test_auto_whitelist_detects_without_a_pool(self):
         args = make_args(barcode_whitelist=[AUTO_BARCODES])
-        isoquant._resolve_barcode_correction(args)
+        options.resolve_barcode_correction(args)
         assert args.detect_cell_barcodes is True
         assert args.n_cells == AUTO_BARCODES
 
     def test_detect_forces_detection_without_n_cells(self, tmp_path):
         args = make_args(barcode_whitelist=whitelist_file(tmp_path),
                          barcode_correction=BarcodeCorrectionMethod.detect.name)
-        isoquant._resolve_barcode_correction(args)
+        options.resolve_barcode_correction(args)
         assert args.detect_cell_barcodes is True
         assert args.n_cells == AUTO_BARCODES
 
@@ -76,7 +76,7 @@ class TestDetectionIsRequested:
         args = make_args(mode="curio", n_cells="5000",
                          barcode_whitelist=whitelist_file(tmp_path))
         with pytest.raises(SystemExit) as excinfo:
-            isoquant._resolve_barcode_correction(args)
+            options.resolve_barcode_correction(args)
         assert excinfo.value.code != 0
 
 
@@ -85,24 +85,57 @@ class TestNCellsIgnoredWarnings:
 
     def test_barcoded_bam_warns(self, caplog):
         args = make_args(n_cells="5000", barcoded_bam=True)
-        isoquant._resolve_barcode_correction(args)
+        options.resolve_barcode_correction(args)
         assert any("--n_cells is ignored" in r.message for r in caplog.records)
         assert args.detect_cell_barcodes is False
 
     def test_barcoded_reads_warns(self, caplog):
         args = make_args(n_cells="5000", barcoded_reads=["reads.tsv"])
-        isoquant._resolve_barcode_correction(args)
+        options.resolve_barcode_correction(args)
         assert any("--n_cells is ignored" in r.message for r in caplog.records)
         assert args.detect_cell_barcodes is False
 
     def test_explicit_whitelist_method_warns(self, caplog, tmp_path):
         args = make_args(n_cells="5000", barcode_whitelist=whitelist_file(tmp_path),
                          barcode_correction=BarcodeCorrectionMethod.whitelist.name)
-        isoquant._resolve_barcode_correction(args)
+        options.resolve_barcode_correction(args)
         assert any("--n_cells is ignored" in r.message for r in caplog.records)
         assert args.detect_cell_barcodes is False
 
     def test_no_warning_when_n_cells_is_honoured(self, caplog, tmp_path):
         args = make_args(n_cells="5000", barcode_whitelist=whitelist_file(tmp_path))
-        isoquant._resolve_barcode_correction(args)
+        options.resolve_barcode_correction(args)
         assert not any("--n_cells is ignored" in r.message for r in caplog.records)
+
+
+class TestBothEntryPointsAgree:
+    """isoquant.py and isoquant_detect_barcodes.py used to carry separate copies of this."""
+
+    @staticmethod
+    def standalone_args(**kwargs):
+        """The standalone tool names the whitelist --barcodes and has no --barcoded_*."""
+        args = make_args(**kwargs)
+        args.barcodes = args.barcode_whitelist
+        del args.barcode_whitelist
+        del args.barcoded_reads
+        del args.barcoded_bam
+        return args
+
+    @pytest.mark.parametrize("kwargs", [
+        {},
+        {"n_cells": "5000"},
+        {"n_cells": "auto"},
+        {"barcode_whitelist": [AUTO_BARCODES]},
+        {"barcode_correction": BarcodeCorrectionMethod.detect.name},
+    ])
+    def test_same_decision_under_either_option_name(self, kwargs, tmp_path):
+        if "barcode_whitelist" not in kwargs:
+            kwargs = dict(kwargs, barcode_whitelist=whitelist_file(tmp_path))
+        pipeline_args = make_args(**kwargs)
+        standalone = self.standalone_args(**kwargs)
+
+        options.resolve_barcode_correction(pipeline_args)
+        options.resolve_barcode_correction(standalone, whitelist_option="--barcodes")
+
+        assert pipeline_args.detect_cell_barcodes == standalone.detect_cell_barcodes
+        assert pipeline_args.n_cells == standalone.n_cells
